+++ backend/db/migrations/000031_add_https_supported.down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE devices DROP COLUMN https_supported;